\name{demoS3imports-package}
\alias{demoS3imports-package}
\alias{demoS3imports}
\docType{package}
\title{
\packageTitle{demoS3imports}
}
\description{
\packageDescription{demoS3imports}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{demoS3imports}
\packageIndices{demoS3imports}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{demoS3imports}

Maintainer: \packageMaintainer{demoS3imports}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
